requirements.txt
aniso8601==4.0.1
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.1
coveralls==1.5.1
cryptography==2.3.1
docopt==0.6.2
Flask==1.0.2
Flask-RESTful==0.3.6
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
psycopg2-binary==2.7.6
py==1.7.0
py3dns==3.2.0
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pytest==3.10.0
pytest-cov==2.6.0
pytz==2018.7
PyYAML==3.13
requests==2.20.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11
